requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.3.1
asn1crypto==0.24.0
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
Django==3.1.3
django-auth-ldap==2.2.0
django-cuser==2017.3.16
django-environ==0.4.5
django-jazzmin==2.4.8
django-modeltranslation==0.17.3
django-rest-auth==0.9.5
django-tinymce==3.4.0
djangorestframework==3.12.2
drf-yasg==1.20.0
filelock==3.0.12
httpie==0.9.8
httplib2==0.9.2
idna==2.6
importlib-metadata==2.0.0
importlib-resources==3.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==2.0.1
modeltranslation==0.25
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
openshot-qt==2.4.1
packaging==21.3
pexpect==4.2.1
Pillow==5.1.0
pipenv==2020.11.15
protobuf==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.2.0
PyGObject==3.26.1
pyldap==3.0.0.post1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyparsing==3.0.6
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.6.1
python-debian==0.1.32
python-ldap==3.3.1
pytz==2020.4
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.2
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scour==0.36
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
sqlparse==0.4.1
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==20.2.0
virtualenv-clone==0.5.4
wadllib==1.3.2
xkit==0.0.0
zipp==3.4.0
zope.interface==4.3.2
